refactor(protocol): rename negotiation.insights.generator.ts to insight.generator.ts (IND-297)
