+++ packages/protocol/src/index.ts
@@ -83,8 +83,8 @@ export { generateInviteMessage } from "./contact/contact.inviter.js";
 export type { InviteInput, InviteOutput } from "./contact/contact.inviter.js";
 export { IntentIndexer } from "./intent/intent.indexer.js";
 export { LensInferrer } from "./shared/hyde/lens.inferrer.js";
-export { NegotiationInsightsGenerator } from "./negotiation/negotiation.insights.generator.js";
-export type { NegotiationDigest } from "./negotiation/negotiation.insights.generator.js";
+export { NegotiationInsightsGenerator } from "./negotiation/insight.generator.js";
+export type { NegotiationDigest } from "./negotiation/insight.generator.js";
 export { IndexNegotiator } from "./negotiation/negotiation.agent.js";
 export type { NegotiationAgentInput } from "./negotiation/negotiation.agent.js";
 export { OpportunityEvaluator } from "./opportunity/opportunity.evaluator.js";
