+++ packages/protocol/src/negotiation/insight.generator.ts
